Add service class definition validator
