--- websocket-compiler-plugin-tests/src/test/resources/ballerina_sources/sample_package_4/CompilerPlugin.toml
@@ -1,6 +1,0 @@
-[plugin]
-id = "sample-package"
-class = "io.ballerina.stdlib.websocket.plugin.WebSocketCompilerPlugin"
-
-[[dependency]]
-path = "../websocket-compiler-plugin/build/libs/websocket-compiler-plugin-@project.version@.jar"
